pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bankfragility"
version = "0.1.0"
description = "Public-data bank deposit stickiness, ALM mismatch, and fragility index pipeline"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.12",
    "ffiec-data-collector>=2.0.0rc1",
    "lxml>=5.3",
    "numpy>=2.1",
    "pandas>=2.2",
    "pyarrow>=18.0",
    "pyyaml>=6.0",
    "requests>=2.32",
    "scikit-learn>=1.5",
    "lightgbm>=4.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.3"]

[tool.setuptools.packages.find]
where = ["src"]
